+++ criu/Makefile.crtools
@@ -49,6 +49,7 @@ obj-y			+= page-pipe.o
 obj-y			+= pagemap.o
 obj-y			+= page-xfer.o
 obj-y			+= parasite-syscall.o
+obj-y			+= sigacts-dump.o
 obj-y			+= pie-util.o
 obj-y			+= pipes.o
 obj-y			+= plugin.o
